[dhctl] moved it to the secretedit function for test coverage
